+++ src/js/components/containers/IndexPage/IndexPage.container.jsx
@@ -1,5 +1,5 @@
 import React, { Fragment } from 'react';
-import { indexPageView as IndexPageView } from 'components/indexPage';
+import { indexPageView as IndexPageView } from 'views/indexPage';
 
 
 export const IndexPage = () => {
