+++ src/js/components/views/footer/footer.style.scss
@@ -11,6 +11,7 @@
     align-items: center;
     justify-content: space-evenly;
     border-top: 0.5px solid $standard-color-white;
+    max-width: 100vw;
 
     &__message {
         font-size: $theme-font-size-6;
