+++ src/js/components/views/footer/footer.view.jsx
