+++ src/js/components/views/footer/index.js
