+++ src/js/components/views/header/header.style.scss
@@ -6,17 +6,17 @@
     height: $header-height;
     background: black;
     color: white;
-    padding: 0 15vw;
+    // padding: 0 15vw;
     display: flex;
     align-items: center;
     border-bottom: 0.5px solid $standard-color-white;
     max-width: 100vw;
 
     &__logo {
+        max-width: 100%;
         height: 100%;
         flex: 1;
         font-size: $theme-font-size-8;
-        width: 100%;
         display: flex;
         justify-content: center;
         align-items: center;
@@ -30,6 +30,7 @@
 
     &__wrapper {
         height: 100%;
+        max-width: 100%;
         flex: 4;
         &__navs {
             height: 100%;
