+++ src/js/components/views/indexPage/indexPage.style.scss
