+++ src/js/components/views/indexPage/indexPage.view.jsx
@@ -1,5 +1,7 @@
 import React from 'react';
-import { RouterOutlet } from '../../router';
+// import { RouterOutlet } from '../../../router';
+// import { RouterOutlet } from '../../../router';
+import { RouterOutlet } from '../../../router'
 
 export const indexPageView = () => {
   return (
