+++ src/js/pages/about/about.page.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 export const aboutPage = () => {
   return (
-    <h1>About Page</h1>
+    <div className='ui-about-page' >About Page</div>
   )
 }
