+++ src/js/pages/about/about.style.scss
@@ -3,4 +3,5 @@
 
 .ui-about-page {
     box-sizing: border-box;
+    color: $theme-primary-color-1;
 }
