+++ src/js/pages/contact/contact.style.scss
@@ -3,4 +3,5 @@
 
 .ui-contact-page {
     box-sizing: border-box;
+    color: $theme-primary-color-1;
 }
