+++ src/js/pages/home/home.style.scss
@@ -8,7 +8,7 @@
     display: flex;
     flex-direction: column;
     color: $standard-color-white;
-    padding: 20px 10px;
+    padding: 2rem;
 
     &__left {
         display: flex;
@@ -17,7 +17,7 @@
 
         &__wrapper {
             &__image {
-                width: 35rem;
+                width: calc(100% - 6rem);
                 border-radius: 5%;
             }
         }
@@ -29,18 +29,15 @@
         &__title {
             text-align: center;
             font-size: $theme-font-size-8;
-            color: $theme-primary-color-1;
+            // color: $theme-primary-color-1;
             margin: 10px 0;
         }
 
         &__desc {
             font-size: $theme-font-size-5;
-            letter-spacing: 1px;
-            padding: 8px 20px;
+            letter-spacing: 0.5px;
+            // padding: 8px 20px;
+            width: calc( 100% - 6rem );
         }
-    }
-
-    @media screen {
-        
     }
 }
