+++ src/js/pages/notFound/notFound.style.scss
@@ -3,4 +3,5 @@
 
 .ui-not-found-page {
     box-sizing: border-box;
+    color: $theme-primary-color-1;
 }
