+++ src/js/pages/portfolio/portfolio.style.scss
@@ -3,4 +3,5 @@
 
 .ui-portfolio-page {
     box-sizing: border-box;
+    color: $theme-primary-color-1;
 }
