+++ webpack.config.extra.js
@@ -4,10 +4,10 @@ exports.alias = {
     'constants': path.resolve( __dirname, 'src/js/constants' ),
 
     // ui components
-    'components': path.resolve( __dirname, 'src/js/components' ),
+    'views': path.resolve( __dirname, 'src/js/components/views' ),
     'hoc': path.resolve( __dirname, 'src/js/hoc' ),
     'modules': path.resolve( __dirname, 'src/js/modules' ),
-    'containers': path.resolve( __dirname, 'src/js/containers' ),
+    'containers': path.resolve( __dirname, 'src/js/components/containers' ),
     'pages': path.resolve( __dirname, 'src/js/pages' ),
     'utils': path.resolve( __dirname, 'src/js/utils' ),
